CBP - Business Transaction - Bank Money Restock - QA Fix GetManager - Plug-in Root #2724
