+++ CBP/plugin/stock_transactions/fermat-cbp-plugin-stock-transactions-bank-money-restock-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/stock_transactions/bank_money_restock/developer/bitdubai/version_1/BusinessTransactionBankMoneyRestockPluginRoot.java
@@ -5,6 +5,7 @@ import com.bitdubai.fermat_api.FermatException;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.abstract_classes.AbstractPlugin;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededAddonReference;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.annotations.NeededPluginReference;
+import com.bitdubai.fermat_api.layer.all_definition.common.system.interfaces.FermatManager;
 import com.bitdubai.fermat_api.layer.all_definition.common.system.utils.PluginVersionReference;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DatabaseManagerForDevelopers;
 import com.bitdubai.fermat_api.layer.all_definition.developer.DeveloperDatabase;
@@ -51,8 +52,6 @@ import java.util.UUID;
  * Created by franklin on 16/11/15.
  */
 public class BusinessTransactionBankMoneyRestockPluginRoot extends AbstractPlugin  implements
-        //TODO: Implementar DealsWiths de los modulos BNK y la Wallet CBP
-        BankMoneyRestockManager,
         DatabaseManagerForDevelopers {
 
     public BusinessTransactionBankMoneyRestockPluginRoot() {
@@ -82,13 +81,13 @@ public class BusinessTransactionBankMoneyRestockPluginRoot extends AbstractPlugi
 
     @Override
     public void start() throws CantStartPluginException {
-        stockTransactionBankMoneyRestockManager = new StockTransactionBankMoneyRestockManager(pluginDatabaseSystem, pluginId);
+
         try {
             Database database = pluginDatabaseSystem.openDatabase(pluginId, BussinessTransactionBankMoneyRestockDatabaseConstants.BANK_MONEY_STOCK_DATABASE_NAME);
 
             //Buscar la manera de arrancar el agente solo cuando hayan transacciones diferentes a COMPLETED
             System.out.println("******* Init Bank Money Restock ******");
-            //testRestock();
+
             startMonitorAgent();
 
             database.closeDatabase();
@@ -123,6 +122,11 @@ public class BusinessTransactionBankMoneyRestockPluginRoot extends AbstractPlugi
     }
 
     @Override
+    public FermatManager getManager() {
+        return new StockTransactionBankMoneyRestockManager(pluginDatabaseSystem, pluginId);
+    }
+
+    @Override
     public List<DeveloperDatabaseTable> getDatabaseTableList(DeveloperObjectFactory developerObjectFactory, DeveloperDatabase developerDatabase) {
         BusinessTransactionBankMoneyRestockDeveloperFactory businessTransactionBankMoneyRestockDeveloperFactory = new BusinessTransactionBankMoneyRestockDeveloperFactory(pluginDatabaseSystem, pluginId);
         return businessTransactionBankMoneyRestockDeveloperFactory.getDatabaseTableList(developerObjectFactory);
@@ -141,33 +145,6 @@ public class BusinessTransactionBankMoneyRestockPluginRoot extends AbstractPlugi
         return developerDatabaseTableRecordList;
     }
 
-    @Override
-    public void createTransactionRestock(String publicKeyActor, FiatCurrency fiatCurrency, String cbpWalletPublicKey, String bankWalletPublicKey, String bankAccount, float amount, String memo, float priceReference, OriginTransaction originTransaction) throws CantCreateBankMoneyRestockException {
-        java.util.Date date = new java.util.Date();
-        Timestamp timestamp = new Timestamp(date.getTime());
-        BankMoneyRestockTransactionImpl bankMoneyRestockTransaction = new BankMoneyRestockTransactionImpl(
-                UUID.randomUUID(),
-                publicKeyActor,
-                fiatCurrency,
-                cbpWalletPublicKey,
-                bankWalletPublicKey,
-                memo,
-                "INIT TRANSACTION",
-                bankAccount,
-                amount,
-                timestamp,
-                TransactionStatusRestockDestock.INIT_TRANSACTION,
-                priceReference,
-                originTransaction);
-
-        try {
-            stockTransactionBankMoneyRestockManager.saveBankMoneyRestockTransactionData(bankMoneyRestockTransaction);
-        } catch (DatabaseOperationException e) {
-            e.printStackTrace();
-        } catch (MissingBankMoneyRestockDataException e) {
-            e.printStackTrace();
-        }
-    }
 
     private BusinessTransactionBankMoneyRestockMonitorAgent businessTransactionBankMoneyRestockMonitorAgent;
     /**
@@ -187,11 +164,4 @@ public class BusinessTransactionBankMoneyRestockPluginRoot extends AbstractPlugi
         }else businessTransactionBankMoneyRestockMonitorAgent.start();
     }
 
-    private void testRestock(){
-        try {
-            createTransactionRestock("publicKeyActor", FiatCurrency.VENEZUELAN_BOLIVAR, "cbpWalletPublicKey", "bnkWalletPublicKey", "account", 250, "memo", 250, OriginTransaction.STOCK_INITIAL);
-        } catch (CantCreateBankMoneyRestockException e) {
-            e.printStackTrace();
-        }
-    }
 }
